requirements.txt
charset-normalizer==2.1.1
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.28.1
vk-api==11.9.9
